Task013. admin. Rename class

+++ app/Http/Controllers/Admin/AnimalsController.php
@@ -4,89 +4,52 @@ namespace App\Http\Controllers\Admin;
 
 use App\Animal;
 use App\Http\Requests\StoreAnimalRequest;
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
 class AnimalsController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         $animals = Animal::paginate(5);
         return view('admin.animals.index', compact('animals'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
+        $this->authorize('create', Animal::class);
         return view('admin.animals.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(StoreAnimalRequest $request)
     {
+        $this->authorize('create', Animal::class);
         Animal::create($request->all());
         return redirect()->route('admin.animals.index')->with(['message' => 'Pet added successfully']);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
+        $this->authorize('update', Animal::class);
         $animals = Animal::findOrFail($id);
         return view('admin.animals.edit', compact('animals'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(StoreAnimalRequest $request, $id)
     {
+        $this->authorize('update', Animal::class);
         $animals = Animal::findOrFail($id);
         $animals->update($request->all());
         return redirect()->route('admin.animals.index')->with(['message' => 'Pet updated successfully']);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
+        $this->authorize('delete', Animal::class);
         $animals = Animal::findOrFail($id);
         $animals->delete();
         return redirect()->route('admin.animals.index')->with(['message' => 'Pet deleted successfully']);
